delete confirmation working
